get_configuration: avoid an unnecessary allocation
"local".into() was allocating a String, but we only need a &str, so Cow
is a good fit here.

This being a one-off thing, I probably wouldn’t actually do this most of
the time since it’s potentially just a distraction from the logic of the
method, but Cow is a useful thing to know about.

In this instance it actually ends up a surprisingly painless alteration,
but I still probably wouldn’t do it.

Source: https://github.com/msupply-foundation/remote-server/commit/b3c746319cdaadf65f52a247afb8965ddc569ff3