Remove unnecessary cloning
You can see from Mutations::insert_requisition, which needs to reuse id
later and so *does* need the clone call, that this is imperfect. But I
think it’s generally better to only clone when you know you need to,
rather than being in the habit of cloning always. It does, however, mean
that some sorts of code mutations are more likely to end up with larger
diffs (when you suddenly need to insert a clone() call earlier in a
method, because you want to use that value again now).