RequisitionRowType ↔ RequisitionType conversions
I’d say that the goal should be to merge these two types, but for now,
I think it’s better to implement From and use that for conversions,
rather than the match block in Mutations::insert_requisitions.
Separating logic from implementation details, and all that.

(At this time, the conversion is only done in one direction, but I’ve
provided both From implementations. Could remove the unused one, or
could keep it in case it’s desired later. Much of a muchness.)