Remove get_configuration unneeded .map_err
It looks like the problem here is that `Config` cannot be
directly deserialized to `Result<Settings, SettingsError>`, as
`SettingsError` does not implement `Deserialize`. This is a tidy-ish
way of doing it, without getting into the messy territory of manually
implementing `Deserialize` for `SettingsError` (which wraps a bunch of
errors which are defined in other crates).

Source: https://github.com/msupply-foundation/remote-server/commit/dab66ecd42f53df6500eeffacc5ce45fdbb65a9a