Add formatters to columns seperate from an accessor
Adding optional formatters to cells separate from the accessor is a little redundant. You could just return a formatted value from the accessor in the case where you want a formatted value. Separating these is simply just because of a separation of concerns. The accessor can be used to, for example, access deeply nested values. For instance, accessing all the lines of an invoice. The formatter, can then just concentrate on formatting a value. For example prepending a dollar sign or capitilising a status