Add passing of translators into formatters
Unsure about this one, but have made it mandatory when calling formatter() that the translators must also be passed. It makes it a bit  easier to write a quick formatter for your column without having to import more dependencies in. Additionally, it is somewhat rare for formatter() to actually be called. Really, it is only going to be called in quite generic places. If for instance, there is a specific column, then it is most likely that you would have a custom Cell component, in which case you would just do your formatting within the cell itself.

Source: https://github.com/msupply-foundation/openmsupply-client/commit/8e2ff997ef6ae78a7d5e89921255acfb209b1473