Add updates to mock server/workers and app to handle the new backend GQL schema
This is quite a messy and big commit, but at lot easier (for me) to do it in one. We have a new graphql schema from the backend team. Ideally we would want to still use the mock-worker for simple unit tests. The apollo-server is being used to power the demo server. So with using the schema ripped from the backend gql server, i've updated the mock resolvers. A smarter person would just run the demo server from the mock worker - though it is only a small amount of work to have the apollo server running.

Currently the app is able to query (but at this point, mutations are still broken) from each of the three sources.