Adjust styles to account for both horizontal and vertical scroll
This change removes the use of an explicit height for the table as it turned out to be not needed. I also added some styling so that horizontal scroll is reenabled which I accidentally prevented in the previous commit.

This commit also makes the header non-sticky, which is a shame. I'll come back to fix that in a new commit as I've spent enough time trying to figure out this flex with scroll!

Source: https://github.com/msupply-foundation/openmsupply-client/commit/04016bcffb9b8aa40e7acebf14d6ae2a640614d1