Start server om "Stage0" config mode when sync config is missing
There are two separate servers now; the normal server and a "stage0"
server which only contains config endpoints. The server can be restarted
through the api to, e.g., use the new config.

Currently, the config file has precedence over the settings in the DB.
To user the DB settings remove the sync entry from the config file.