Update stocktake structured errors
`AdjustmentReasonNotValid` is now a structured error, although we can probably protect against this in front end, it's a bit hard to do and technically speaking this could also occur without front end knowing (due to sync, i.e. if you change reason to inactive and sync while editing stocktake lines)

`StockLinesReducedBelowZero` to return array of `StockLineReducedBelowZero`. If error occurs during finalisation, it is displayed in main table and when we look at batch table it should also be displayed (just through opening it), since batch table needs to display errors based on StockLine (since it shows draft lines, and they may not be inserted yet), it's good to use one common error for each stocktake line, thus collection of `StockLineReducedBelowZero` errors that can be mapped to each stocktakeline.