Log only for initialisation with boolean
I'd love to pass logger as an Option<&mut SyncLogger>, but I can't get it to work without transferring ownership of the logger when I check if it exists with match statements.

I'm sure there is an easy solution, but I can't figure it out.

This is a jerry-rigged solution by passing an additional boolean which removes the need to check if logger is being passed.

Source: https://github.com/msupply-foundation/open-msupply/commit/38032542d9dbd3eaab08c1317a6b98ffc581a48d