Add stocktake detail view
Note haven't changed the button shrinking on the detail tab.
I wonder if we should have the same shrink threshold for simplified mobile view as per shrinkthreshold in these buttons. That would make sense for me?

I think if there is consensus on that, it would be a lot easier both to implement and manage than adding a custom override