Add simplified UI (#7830)
* Add simplified UI for stock list

* Add simplified stockline edit view

* Update detail stock view to remove tabs on mobile view

* Add stocktake detail view

Note haven't changed the button shrinking on the detail tab.
I wonder if we should have the same shrink threshold for simplified mobile view as per shrinkthreshold in these buttons. That would make sense for me?

I think if there is consensus on that, it would be a lot easier both to implement and manage than adding a custom override

* Add stock take list view simplified ui

* Hide comment column by default on stocktakelineedit

* Hide popover columns by default

* Add stocktake simplified UI internal orders

* Review changes

* Fix type error

* Add item number for simplified UI

Using Item # shorthand so that it will display easily on smaller screen without wrapping

* Rename to not use 'use' at beginning of variable name

* Rename to Number of rows

---------

Co-authored-by: Carl Smith <5456533+CarlosNZ@users.noreply.github.com>